+++ src/DateFormField.jsx
@@ -16,7 +16,6 @@ import deepcopy from 'lodash/cloneDeep';
 import omitBy from 'lodash/omitBy';
 import isNil from 'lodash/isNil';
 import Formatter from 'uxcore-formatter';
-import addEventListener from 'rc-util/lib/Dom/addEventListener';
 
 const CalendarPanel = {
   month: Calendar.MonthCalendar,
@@ -25,35 +24,6 @@ const CalendarPanel = {
 };
 
 const getMode = props => props.jsxmode || props.mode;
-const getVerticalAlign = (props) => {
-  // jsxVerticalAlign is an internal varible.
-  let align = props.verticalAlign;
-  if (align === undefined) {
-    align = props.jsxVerticalAlign;
-  }
-  return align;
-};
-
-const getMaxWidth = props => props.inputBoxMaxWidth;
-
-const getIEVer = () => {
-  if (window) {
-    const ua = window.navigator.userAgent;
-    const idx = ua.indexOf('MSIE');
-    if (idx > 0) {
-      // "Mozilla/5.0 (compatible; MSIE 10.0; Windows NT 6.1; WOW64;
-      // Trident/6.0; SLCC2; .NET CLR 2.0.50727)"
-      return parseInt(ua.substring(idx + 5, ua.indexOf('.', idx)), 10);
-    }
-    if (ua.match(/Trident\/7\./)) {
-      // "Mozilla/5.0 (Windows NT 6.1; WOW64; Trident/7.0; SLCC2;
-      // .NET CLR 2.0.50727; rv:11.0) like Gecko"
-      return 11;
-    }
-    return 0;
-  }
-  return 0;
-};
 
 const getPropFromArray = (arr, index) => {
   if (arr instanceof Array) {
@@ -85,7 +55,8 @@ class DateFormField extends FormField {
     }
   }
 
-  componentDidUpdate() {
+  componentDidUpdate(...args) {
+    super.componentDidUpdate(...args);
     const { jsxtype, autoMatchWidth } = this.props;
     const mode = getMode(this.props);
     if (jsxtype === 'cascade' && autoMatchWidth && mode === Constants.MODE.EDIT) {